+++ nautilus_core/core/src/uuid.rs
@@ -13,32 +13,42 @@
 //  limitations under the License.
 // -------------------------------------------------------------------------------------------------
 
-use crate::string::into_cstring;
-use std::ffi::CStr;
+use crate::buffer::Buffer36;
 use std::fmt::{Debug, Display, Formatter, Result};
-use std::os::raw::c_char;
 use uuid::Uuid;
 
 #[repr(C)]
 #[derive(Clone, Hash, PartialEq, Eq)]
 pub struct UUID4 {
-    value: Box<String>,
+    value: Buffer36,
 }
 
 impl UUID4 {
     pub fn new() -> UUID4 {
+        let uuid = Uuid::new_v4();
+        let mut buffer: [u8; 36] = [0; 36];
+        buffer.copy_from_slice(uuid.to_string().into_bytes().as_slice());
         UUID4 {
-            value: Box::new(Uuid::new_v4().to_string()),
+            value: Buffer36 {
+                data: buffer,
+                len: 36,
+            },
         }
     }
 
+    pub fn from_bytes(value: Buffer36) -> UUID4 {
+        UUID4 { value }
+    }
+
     pub fn from_str(s: &str) -> UUID4 {
+        let uuid = Uuid::parse_str(s).unwrap();
+        let mut buffer: [u8; 36] = [0; 36];
+        buffer.copy_from_slice(uuid.to_string().into_bytes().as_slice());
         UUID4 {
-            value: Box::new(
-                Uuid::parse_str(s)
-                    .expect("Invalid UUID4 string")
-                    .to_string(),
-            ),
+            value: Buffer36 {
+                data: buffer,
+                len: 36,
+            },
         }
     }
 }
@@ -51,13 +61,13 @@ impl Default for UUID4 {
 
 impl Debug for UUID4 {
     fn fmt(&self, f: &mut Formatter<'_>) -> Result {
-        write!(f, "{}", self.value)
+        write!(f, "{}", String::from_utf8_lossy(self.value.data.as_slice()))
     }
 }
 
 impl Display for UUID4 {
     fn fmt(&self, f: &mut Formatter<'_>) -> Result {
-        write!(f, "{}", self.value)
+        write!(f, "{}", String::from_utf8_lossy(self.value.data.as_slice()))
     }
 }
 
@@ -76,15 +86,13 @@ pub extern "C" fn uuid4_free(uuid4: UUID4) {
 
 /// Expects `ptr` to be an array of valid UTF-8 chars with the trailing nul byte terminator.
 #[no_mangle]
-pub unsafe extern "C" fn uuid4_from_cstring(ptr: *const c_char) -> UUID4 {
-    // SAFETY: Wraps and checks raw C string `ptr`, then converts to owned `String`
-    let s = CStr::from_ptr(ptr).to_str().expect("invalid C string");
-    UUID4::from_str(s)
+pub extern "C" fn uuid4_from_bytes(value: Buffer36) -> UUID4 {
+    UUID4::from_bytes(value)
 }
 
 #[no_mangle]
-pub extern "C" fn uuid4_to_cstring(uuid: &UUID4) -> *const c_char {
-    into_cstring(uuid.to_string())
+pub extern "C" fn uuid4_to_bytes(uuid: &UUID4) -> Buffer36 {
+    uuid.value.clone()
 }
 
 ////////////////////////////////////////////////////////////////////////////////
@@ -92,12 +100,12 @@ pub extern "C" fn uuid4_to_cstring(uuid: &UUID4) -> *const c_char {
 ////////////////////////////////////////////////////////////////////////////////
 #[cfg(test)]
 mod tests {
-    use crate::uuid::{uuid4_from_cstring, uuid4_new, uuid4_to_cstring, UUID4};
-    use std::ffi::{CStr, CString};
+    use crate::buffer::Buffer36;
+    use crate::uuid::{uuid4_from_bytes, uuid4_new, UUID4};
 
     #[test]
     fn test_new() {
-        let uuid = UUID4::new();
+        let uuid = UUID4::from_str("2d89666b-1a1e-4a75-b193-4eb3b454c757");
 
         assert_eq!(uuid.to_string().len(), 36)
     }
@@ -118,22 +126,20 @@ mod tests {
     }
 
     #[test]
-    fn test_uuid4_from_cstring() {
-        unsafe {
-            let cstring = CString::new("2d89666b-1a1e-4a75-b193-4eb3b454c757").unwrap();
-            let uuid = uuid4_from_cstring(cstring.as_ptr());
+    fn test_uuid4_from_bytes() {
+        let mut buffer: [u8; 36] = [0; 36];
+        buffer.copy_from_slice(
+            "2d89666b-1a1e-4a75-b193-4eb3b454c757"
+                .to_string()
+                .into_bytes()
+                .as_slice(),
+        );
+        let value = Buffer36 {
+            data: buffer,
+            len: 36,
+        };
+        let uuid = uuid4_from_bytes(value);
 
-            assert_eq!(uuid.to_string(), "2d89666b-1a1e-4a75-b193-4eb3b454c757")
-        }
-    }
-
-    #[test]
-    fn test_uuid4_to_cstring() {
-        unsafe {
-            let uuid = UUID4::new();
-            let ptr = uuid4_to_cstring(&uuid);
-
-            assert_eq!(CStr::from_ptr(ptr).to_str().unwrap().len(), 36)
-        }
+        assert_eq!(uuid.to_string(), "2d89666b-1a1e-4a75-b193-4eb3b454c757")
     }
 }