+++ nautilus_trader/core/uuid.pyx
@@ -16,12 +16,10 @@
 import re
 
 from nautilus_trader.core.correctness cimport Condition
+from nautilus_trader.core.rust.core cimport Buffer36
 from nautilus_trader.core.rust.core cimport uuid4_free
-from nautilus_trader.core.rust.core cimport uuid4_from_cstring
+from nautilus_trader.core.rust.core cimport uuid4_from_bytes
 from nautilus_trader.core.rust.core cimport uuid4_new
-from nautilus_trader.core.rust.core cimport uuid4_to_cstring
-from nautilus_trader.core.string cimport cstring_to_pystr
-from nautilus_trader.core.string cimport pystr_to_cstring
 
 
 _UUID_REGEX = re.compile("[0-F]{8}-([0-F]{4}-){3}[0-F]{12}", re.I)
@@ -51,32 +49,40 @@ cdef class UUID4:
         if value is None:
             # Create a new UUID4 from Rust
             self._uuid4 = uuid4_new()  # `UUID4_t` owned from Rust
-            self.value = cstring_to_pystr(uuid4_to_cstring(&self._uuid4))
         else:
             Condition.true(_UUID_REGEX.match(value), "value is not a valid UUID")
             self._uuid4 = self._uuid4_from_pystring(value)
-            self.value = value
 
     cdef UUID4_t _uuid4_from_pystring(self, str value) except *:
-        return uuid4_from_cstring(pystr_to_cstring(value))  # `value` moved to Rust, `UUID4_t` owned from Rust
+        cdef Buffer36 buffer
+        buffer.data = value.encode()
+        buffer.len = 36
+        self._uuid4 = uuid4_from_bytes(buffer)
+        return uuid4_from_bytes(buffer)  # `value` moved to Rust, `UUID4_t` owned from Rust
+
+    cpdef str to_str(self):
+        return self._uuid4.value.data[:36].decode()
 
     def __getstate__(self):
-        return self.value
+        return self._uuid4.value.data[:36]
 
     def __setstate__(self, state):
-        self._uuid4 = self._uuid4_from_pystring(state)
+        cdef Buffer36 buffer
+        buffer.data = state
+        buffer.len = 36
+        self._uuid4 = uuid4_from_bytes(buffer)
 
     def __eq__(self, UUID4 other) -> bool:
-        return self.value == other.value
+        return self._uuid4.value.data[:36] == other._uuid4.value.data[:36]
 
     def __hash__(self) -> int:
-        return hash(self.value)
+        return hash(self._uuid4.value.data)
 
     def __str__(self) -> str:
-        return self.value
+        return self.to_str()
 
     def __repr__(self) -> str:
-        return f"{type(self).__name__}('{self.value}')"
+        return f"{type(self).__name__}('{self.to_str()}')"
 
     def __del__(self) -> None:
         uuid4_free(self._uuid4)  # `self._uuid4` moved to Rust (then dropped)